[Code]optimized code; Support resume training

+++ utils/train_utils.py
@@ -1,13 +1,70 @@
 import os
 import torch
+import torch.distributed as dist
 import sys
 import types
 import importlib
+import random
 from tqdm import tqdm
+import numpy as np
 from numpy import mean
 from terminaltables import AsciiTable
-from configs import backbones
+from torch.optim import Optimizer
 from core.evaluations import evaluate
+from utils.checkpoint import save_checkpoint,load_checkpoint
+from utils.common import get_dist_info
+
+
+def init_random_seed(seed=None, device='cuda'):
+    """Initialize random seed.
+
+    If the seed is not set, the seed will be automatically randomized,
+    and then broadcast to all processes to prevent some potential bugs.
+
+    Args:
+        seed (int, Optional): The seed. Default to None.
+        device (str): The device where the seed will be put on.
+            Default to 'cuda'.
+
+    Returns:
+        int: Seed to be used.
+    """
+    if seed is not None:
+        return seed
+
+    # Make sure all ranks share the same random seed to prevent
+    # some potential bugs. Please refer to
+    # https://github.com/open-mmlab/mmdetection/issues/6339
+    rank, world_size = get_dist_info()
+    seed = np.random.randint(2**31)
+    if world_size == 1:
+        return seed
+
+    if rank == 0:
+        random_num = torch.tensor(seed, dtype=torch.int32, device=device)
+    else:
+        random_num = torch.tensor(0, dtype=torch.int32, device=device)
+    dist.broadcast(random_num, src=0)
+    return random_num.item()
+
+
+def set_random_seed(seed, deterministic=False):
+    """Set random seed.
+
+    Args:
+        seed (int): Seed to be used.
+        deterministic (bool): Whether to set the deterministic option for
+            CUDNN backend, i.e., set `torch.backends.cudnn.deterministic`
+            to True and `torch.backends.cudnn.benchmark` to False.
+            Default: False.
+    """
+    random.seed(seed)
+    np.random.seed(seed)
+    torch.manual_seed(seed)
+    torch.cuda.manual_seed_all(seed)
+    if deterministic:
+        torch.backends.cudnn.deterministic = True
+        torch.backends.cudnn.benchmark = False
 
 '''
 读取配置文件
@@ -82,10 +139,68 @@ def get_lr(optimizer):
         return param_group['lr']
 
 '''
+从中断训练处恢复训练
+'''
+def resume_model(model, runner, checkpoint, meta, resume_optimizer=True, map_location='default'):
+    if map_location == 'default':
+        if torch.cuda.is_available():
+            device_id = torch.cuda.current_device()
+            checkpoint = load_checkpoint(
+                model,
+                checkpoint,
+                map_location=lambda storage, loc: storage.cuda(device_id))
+        else:
+            checkpoint = load_checkpoint(model, checkpoint)
+    else:
+        checkpoint = load_checkpoint(
+            model, checkpoint, map_location=map_location)
+
+    runner['epoch'] = checkpoint['meta']['epoch']
+    runner['iter'] = checkpoint['meta']['iter']
+    runner['best_train_weight'] = checkpoint['meta']['best_train_weight']
+    runner['last_weight'] = checkpoint['meta']['last_weight']
+    runner['best_val_weight'] = checkpoint['meta']['best_val_weight']
+    runner['best_train_loss'] = checkpoint['meta']['best_train_loss']
+    runner['best_val_acc'] = checkpoint['meta']['best_val_acc']
+    if meta is None:
+        meta = {}
+
+    # # Re-calculate the number of iterations when resuming
+    # # models with different number of GPUs
+    # if 'config' in checkpoint['meta']:
+    #     config = mmcv.Config.fromstring(
+    #         checkpoint['meta']['config'], file_format='.py')
+    #     previous_gpu_ids = config.get('gpu_ids', None)
+    #     if previous_gpu_ids and len(previous_gpu_ids) > 0 and len(
+    #             previous_gpu_ids) != self.world_size:
+    #         self._iter = int(self._iter * len(previous_gpu_ids) /
+    #                             self.world_size)
+
+    # resume meta information meta
+    meta = checkpoint['meta']
+
+    if 'optimizer' in checkpoint and resume_optimizer:
+        if isinstance(runner['optimizer'], Optimizer):
+            runner['optimizer'].load_state_dict(checkpoint['optimizer'])
+        elif isinstance(runner['optimizer'], dict):
+            for k in runner['optimizer'].keys():
+                runner.optimizer[k].load_state_dict(
+                    checkpoint['optimizer'][k])
+        else:
+            raise TypeError(
+                'Optimizer should be dict or torch.optim.Optimizer '
+                f'but got {type(runner.optimizer)}')
+
+    print('resumed epoch %d, iter %d'% (runner['epoch'], runner['iter']))
+    return model, runner, meta
+
+'''
 训练
 '''
-def train(model, runner, lr_update_func, device, epoch, epoches, save_dir, train_history):
+def train(model, runner, lr_update_func, device, epoch, epoches, meta):
     train_loss = 0
+    runner['epoch'] = epoch + 1
+    meta['epoch'] = runner['epoch']
     
     model.train()
     with tqdm(total=len(runner.get('train_loader')),desc=f'Train: Epoch {epoch + 1}/{epoches}',postfix=dict,mininterval=0.3) as pbar:
@@ -106,32 +221,28 @@ def train(model, runner, lr_update_func, device, epoch, epoches, save_dir, train
                                 'Lr' : get_lr(runner.get('optimizer'))
                                 })
             runner['iter'] += 1
+            meta['iter'] = runner['iter']
             pbar.update(1)
     
-    train_history.update(train_loss / (iter + 1),'train')
+    meta['train_info']['train_loss'].append(train_loss / (iter + 1))
             
     if train_loss/len(runner.get('train_loader')) < runner.get('best_train_loss') :
         runner['best_train_loss'] = train_loss/len(runner.get('train_loader'))
-        if epoch > 0:
+        meta['best_train_loss'] = runner['best_train_loss']
+        if epoch > 0 and os.path.isfile(runner['best_train_weight']):
             os.remove(runner['best_train_weight'])
-        runner['best_train_weight'] = os.path.join(save_dir,'Train_Epoch{:03}-Loss{:.3f}.pth'.format(epoch+1,train_loss / len(runner.get('train_loader'))))
-        torch.save(model.state_dict(),runner.get('best_train_weight'))
-    
-    if epoch > 0:
-        os.remove(runner['last_weight'])
-    runner['last_weight'] = os.path.join(save_dir,'Last_Epoch{:03}.pth'.format(epoch+1))
-    torch.save(model.state_dict(),runner.get('last_weight'))
+        runner['best_train_weight'] = os.path.join(meta['save_dir'],'Train_Epoch{:03}-Loss{:.3f}.pth'.format(epoch+1,train_loss / len(runner.get('train_loader'))))
+        meta['best_train_weight'] = runner['best_train_weight']
+        save_checkpoint(model,runner.get('best_train_weight'),runner.get('optimizer'),meta)
     
 
-def validation(model, runner, cfg, device, epoch, epoches, save_dir, train_history):
-
+def validation(model, runner, cfg, device, epoch, epoches, meta):
     preds,targets = [],[]
     model.eval()
     with torch.no_grad():
         with tqdm(total=len(runner.get('val_loader')), desc=f'Test : Epoch {epoch + 1}/{epoches}',mininterval=0.3) as pbar:
             for iter, batch in enumerate(runner.get('val_loader')):
                 images, target, _ = batch
-                #runner.get('optimizer').zero_grad()
                 outputs = model(images.to(device),return_loss=False)
                 preds.append(outputs)
                 targets.append(target.to(device))   
@@ -139,7 +250,7 @@ def validation(model, runner, cfg, device, epoch, epoches, save_dir, train_histo
                 
     eval_results = evaluate(torch.cat(preds),torch.cat(targets),cfg.get('metrics'),cfg.get('metric_options'))
     
-    train_history.update(eval_results,'test')
+    meta['train_info']['val_acc'].append(eval_results)
     
     TITLE = 'Validation Results'
     TABLE_DATA = (
@@ -153,9 +264,17 @@ def validation(model, runner, cfg, device, epoch, epoches, save_dir, train_histo
     print(table_instance.table)
     print()
              
-    if save_dir and eval_results.get('accuracy_top-1') > runner.get('best_val_acc'):
+    if eval_results.get('accuracy_top-1') > runner.get('best_val_acc'):
         runner['best_val_acc'] = eval_results.get('accuracy_top-1')
-        if epoch > 0:
+        meta['best_val_acc'] = runner['best_val_acc']
+        if epoch > 0 and os.path.isfile(runner['best_val_weight']):
             os.remove(runner['best_val_weight'])
-        runner['best_val_weight'] = os.path.join(save_dir,'Val_Epoch{:03}-Acc{:.3f}.pth'.format(epoch+1,eval_results.get('accuracy_top-1')))
-        torch.save(model.state_dict(),runner.get('best_val_weight'))
+        runner['best_val_weight'] = os.path.join(meta['save_dir'],'Val_Epoch{:03}-Acc{:.3f}.pth'.format(epoch+1,eval_results.get('accuracy_top-1')))
+        meta['best_val_weight'] = runner['best_val_weight']
+        save_checkpoint(model,runner.get('best_val_weight'),runner.get('optimizer'),meta)
+
+    if epoch > 0 and os.path.isfile(runner['last_weight']):
+        os.remove(runner['last_weight'])
+    runner['last_weight'] = os.path.join(meta['save_dir'],'Last_Epoch{:03}.pth'.format(epoch+1))
+    meta['last_weight'] = runner['last_weight']
+    save_checkpoint(model,runner.get('last_weight'),runner.get('optimizer'),meta)
